reafactor: MEMBER_LIST model로 파일로 분류
