--- src/model/gnb/index.tsx
@@ -1,14 +1,0 @@
-const GNB_LIST = [
-  {
-    title: '백오피스 홈',
-    path: '',
-    key: 'back_home',
-  },
-  {
-    title: '회원관리',
-    path: 'member/manage',
-    key: 'back_manage',
-  },
-]
-
-export { GNB_LIST }
