+++ src/model/index.tsx
@@ -1,0 +1,20 @@
+const GNB_LIST = [
+  {
+    title: '백오피스 홈',
+    path: '',
+    key: 'back_home',
+  },
+  {
+    title: '회원관리',
+    path: 'member/manage',
+    key: 'back_manage',
+  },
+]
+
+const MEMBER_LIST = [
+  { member_seq: '136', crt_ymdt: '2022-02-26 12:40:14', id: 'yhorong21' },
+  { member_seq: '328', crt_ymdt: '2022-04-16 07:19:31', id: 'seoltosil95' },
+  { member_seq: '380', crt_ymdt: '2022-04-17 03:00:23', id: 'yujaemin92' },
+]
+
+export { GNB_LIST, MEMBER_LIST }
