+++ src/routes/ManagePage/state.ts
@@ -1,3 +1,4 @@
+import { MEMBER_LIST } from 'model'
 import { atom } from 'recoil'
 
 interface IMember {
@@ -5,12 +6,6 @@ interface IMember {
   crt_ymdt: string
   id: string
 }
-
-export const MEMBER_LIST = [
-  { member_seq: '136', crt_ymdt: '2022-02-26 12:40:14', id: 'yhorong21' },
-  { member_seq: '328', crt_ymdt: '2022-04-16 07:19:31', id: 'seoltosil95' },
-  { member_seq: '380', crt_ymdt: '2022-04-17 03:00:23', id: 'yujaemin92' },
-]
 
 export const filteredListState = atom<IMember[]>({
   key: '#filteredListState',
